🧩 4GSCYN task: satisfy hotspot contract
