+++ packages/agentplane/src/commands/shared/workflow-step-branch-state.ts
@@ -1,4 +1,21 @@
-import type { WorkflowRouteState } from "./workflow-step.js";
+import type { WorkflowRouteState, WorkflowStep } from "./workflow-step.js";
+import { terminalStep } from "./workflow-step-factory.js";
+
+export function blockedTaskStep(state: WorkflowRouteState): WorkflowStep {
+  return terminalStep({
+    state,
+    id: "terminal.task_blocked",
+    code: "task_blocked",
+    phase: "task_blocked",
+    checkout: "task_worktree",
+    role: "CODER",
+    outcome: "attention_required",
+    summary:
+      "the task is blocked by a recorded semantic outcome; resolve the blocker and explicitly resume the task before requesting another implementation episode",
+    evidenceMissing: ["blocker_resolution"],
+    selectedBlocker: null,
+  });
+}
 
 export function primaryIncludeTaskIds(state: WorkflowRouteState): readonly string[] {
   return state.batchOwnership.role === "primary" ? [...state.batchOwnership.includedTaskIds] : [];
