+++ packages/agentplane/src/commands/shared/workflow-step-branch.ts
@@ -1,7 +1,11 @@
 import type { WorkflowRouteState, WorkflowStep } from "./workflow-step.js";
 import type { RouteBlocker } from "./route-oracle.js";
 import { conflictReworkRouteStep } from "./workflow-step-conflict-rework.js";
-import { preMergeCommit, primaryIncludeTaskIds } from "./workflow-step-branch-state.js";
+import {
+  blockedTaskStep,
+  preMergeCommit,
+  primaryIncludeTaskIds,
+} from "./workflow-step-branch-state.js";
 import { supersededProviderConflictStep } from "./workflow-step-provider-conflict-superseded.js";
 import { needsQualityEvidenceRefresh } from "./workflow-step-quality.js";
 import { integrationQueueStep } from "./workflow-step-integration-queue.js";
@@ -384,21 +388,7 @@ export function branchStep(state: WorkflowRouteState): WorkflowStep {
       selectedBlocker: null,
     });
   }
-  if (status === "BLOCKED") {
-    return terminalStep({
-      state,
-      id: "terminal.task_blocked",
-      code: "task_blocked",
-      phase: "task_blocked",
-      checkout: "task_worktree",
-      role: "CODER",
-      outcome: "attention_required",
-      summary:
-        "the task is blocked by a recorded semantic outcome; resolve the blocker and explicitly resume the task before requesting another implementation episode",
-      evidenceMissing: ["blocker_resolution"],
-      selectedBlocker: null,
-    });
-  }
+  if (status === "BLOCKED") return blockedTaskStep(state);
   if (state.batchOwnership.role === "included") return includedBatchStep(state);
   if (!state.prFlow?.branch.name && verifiedIncludedClosureCandidate(state.task)) {
     return cliOperationStep({
